+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # TsdkArc Packages Changelog
 
+
+## tsdkarc-x@0.0.5
+
+- Fix `tsdkarc-x/esm` not build
+
+## tsdkarc-x@0.1.1
+
+- Fix `tsdkarc-x/esm` not build
+
 ## tsdkarc-x@0.1.0
 
 - Breaking⚠️: Refactor the middleware implementation with better DX